refactor to pull apart the gametree
